fix clean branches
