+++ src/lib/git_branch.rs
@@ -62,37 +62,3 @@ impl std::clone::Clone for GitBranch {
         }
     }
 }
-
-pub(crate) fn parse_raw_branch_data(
-    raw_branch_data: &Vec<String>,
-    remote_identifier: &str,
-) -> Vec<GitBranch> {
-    let branch_list = raw_branch_data
-        .iter()
-        .map(|s| match s.starts_with(remote_identifier) {
-            true => {
-                let remote_branch_name = s
-                    .split((remote_identifier.to_string() + "/").as_str())
-                    .nth(1)
-                    .unwrap();
-
-                let remote = remote_branch_name.split("/").nth(0).unwrap().to_string();
-                let branch: String = remote_branch_name
-                    .split("/")
-                    .skip(1)
-                    .collect::<Vec<&str>>()
-                    .join("/");
-
-                GitBranch {
-                    name: branch,
-                    remote_prefix: Some(remote),
-                }
-            }
-            false => GitBranch {
-                name: s.clone(),
-                remote_prefix: None,
-            },
-        }).collect();
-
-    return branch_list;
-}
